api/requirements.txt
Flask
Flask-SQLAlchemy
numpy
visvalingamwyatt
feedparser
beautifulsoup4
requests
flask-jwt-extended